+++ src/PopupsWrapper.tsx
@@ -8,15 +8,20 @@ interface PopupsWrapperProps {
 
 interface SinglePopupLifeCycleProps {
   currentPopup: PopupItem;
-  onClose(guid: string): any;
+  onClose(
+    guid: string,
+    onConsumerOnCloseCallback: Function,
+    ...args: any[]
+  ): any;
   isOpen: boolean;
 }
 
 class SinglePopupLifeCycle extends React.Component<SinglePopupLifeCycleProps> {
   state = { isOpen: false };
+
   constructor(props) {
-      super(props);
-      this.onClose = this.onClose.bind(this)
+    super(props);
+    this.onClose = this.onClose.bind(this);
   }
 
   componentDidMount(): void {
@@ -35,10 +40,18 @@ class SinglePopupLifeCycle extends React.Component<SinglePopupLifeCycleProps> {
 
     return null;
   }
-  private onClose(...params: any[]) {
-      const {currentPopup, onClose} = this.props;
-      onClose(currentPopup.guid);
-      currentPopup.props?.onClose?.(...params);
+
+  private onClose(...closeArgs: any[]) {
+    const { currentPopup, onClose } = this.props;
+    if (currentPopup.props?.onClose) {
+      onClose(
+        currentPopup.guid,
+        () => currentPopup.props?.onClose(...closeArgs),
+        closeArgs,
+      );
+    } else {
+      onClose(currentPopup.guid, undefined, closeArgs);
+    }
   }
 
   render() {
@@ -54,16 +67,21 @@ class SinglePopupLifeCycle extends React.Component<SinglePopupLifeCycleProps> {
 }
 
 export class PopupsWrapper extends React.Component<PopupsWrapperProps> {
-    constructor(props) {
-        super(props);
-        this.onClose = this.onClose.bind(this);
-    }
+  constructor(props) {
+    super(props);
+    this.onClose = this.onClose.bind(this);
+  }
+
   componentDidMount(): void {
     this.props.popupManager.subscribeOnPopupsChange(() => this.forceUpdate());
   }
 
-  private onClose(guid: string) {
-    this.props.popupManager.close(guid);
+  private onClose(
+    guid: string,
+    onConsumerOnCloseCallback?: Function,
+    closedArgs?: any[],
+  ) {
+    this.props.popupManager.close(guid, onConsumerOnCloseCallback, closedArgs);
   }
 
   public render() {