+++ src/popupsDef.ts
@@ -1,13 +1,14 @@
 import { PopupManager } from './popupManager';
 
-export interface popupInstance {
-  close: Function;
-  unmount: Function;
-}
-
 export interface PopupProps {
   onClose?: (...args: any[]) => any;
   isOpen?: boolean;
+}
+
+export interface popupInstance {
+  close: Function;
+  unmount: Function;
+  response: Promise<any>;
 }
 
 export interface WithPopupsProps {